# Conflicts:
#	flickit-assessment-common/src/main/resources/i18n/kit/messages_en.properties
#	flickit-assessment-data/src/main/java/org/flickit/assessment/data/jpa/kit/levelcompetence/LevelCompetenceJpaRepository.java
#	flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/adapter/out/persistence/levelcompetence/LevelCompetencePersistenceJpaAdapter.java
#	flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/application/port/out/levelcomptenece/DeleteLevelCompetencePort.java
#	flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/common/ErrorMessageKey.java
#	flickit-assessment-kit/src/test/resources/postman/flickit-assessment-kit.postman_collection.json

+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/adapter/out/persistence/maturitylevel/MaturityLevelPersistenceJpaAdapter.java
@@ -41,6 +41,8 @@ public class MaturityLevelPersistenceJpaAdapter implements
 
     @Override
     public void delete(Long id, Long kitVersionId) {
+        if(!repository.existsByIdAndKitVersionId(id, kitVersionId))
+            throw new ResourceNotFoundException(MATURITY_LEVEL_ID_NOT_FOUND);
         repository.deleteByIdAndKitVersionId(id, kitVersionId);
     }
